cmd/read: actually implement the read cmd
The original cmd/read.go was committed by mistake and it wasn't implemented.
This commit fixes that. The difference between it and cmd/dump.go is that it reads
data using the Fetch api, instead of fetching the raw data directly from whisper
file. In a way, it is similar to reading data with render api in go-carbon.

An example usage:
    ./read -quiet -from '2021-05-12 08:11:21' -until '2021-05-12 08:12:22'