cwhisper: bug fixes related to ooo writes in the aggregation buffer
using the 6 data points as a buffer example in an archive: | t0 | t1 | t2 | t3 | t4 | t5 |

when there are values in t1, t3, t4. If cwhisper receives update at t0, it would trigger
buffer aggregation (i.e. propagation). this is mostly fine but when t0 belongs to the same
aggregated interval as t1. the propagation is actually premature and can wait.

paring with another bug in archiveInfo.AppendPointsToBlock, where it would append data points
if the it has the same interval/epoch as the last point. This might causes space waste as
in the current cwhisper design, we do not intended to append duplicate and ooo data points.

this commit fixes both issues. the aggregation buffer would tolerate ooo writes when it is
possible and archiveInfo.AppendPointsToBlock would stop appending data with the same epoch.